+++ src/GSettingsDlg.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <Gtk/Window.hpp>
+
+class GFont;
+
+class GSettingsDlg
+{
+public:
+    GSettingsDlg(gir::Gtk::Window, GFont &);
+};