+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow launching Gtk inspector from the application menu
 - Help->About dialog
 - Persistent configuration of guifont via GSettings
+- GUI dialog to configure settings
 
 ### Fixed